+++ src/main/java/de/epsdev/minecord/Minecord.java
@@ -6,14 +6,18 @@ import de.epsdev.minecord.events.e_OnPlayerChat;
 import org.bukkit.plugin.PluginManager;
 import org.bukkit.plugin.java.JavaPlugin;
 
+import java.util.logging.Logger;
+
 public final class Minecord extends JavaPlugin {
 
     public static Bot bot;
     public static PluginConfig pluginConfig;
+    public static Logger pluginLogger;
 
     @Override
     public void onEnable() {
         pluginConfig = new PluginConfig(this);
+        pluginLogger = this.getLogger();
         bot = new Bot();
 
         registerEvents();